corrections after make debug

+++ ft_strnstr.c
@@ -13,7 +13,7 @@ char	*ft_strnstr(const char *haystack, const char *needle, size_t n)
 		j = 0;
 		while (haystack[i + j] == needle [j] && needle[j] != '\0' && j + i < n)
 			j++;
-		if (j == ft_strlen(needle))
+		if (j == (size_t)ft_strlen(needle))
 			return ((char *)&haystack[i]);
 		i++;
 	}
